requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.1
dill==0.3.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
libretranslatepy==2.1.1
loguru==0.5.3
lxml==4.6.3
multidict==5.1.0
multiprocess==0.70.12.2
pathos==0.2.8
pox==0.3.0
ppft==1.6.6.4
pyenchant==3.2.1
PyExecJS==1.5.1
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
translators==4.9.5
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.6.3
